feat(obd2): reliability overhaul (v1.1.26) — batch-desync fix, DTC multi-ECU merge, 29-bit ISO 15765-4
OBD2 audit + overhaul:
- Fix PID 0x14 batch desync: a default-set PID with the wrong response
  length silently dropped every PID after it in a batch (the "values
  don't show up" bug). Add resp_len to the PID table + J1979 table test.
- Batch-reject latch: on NRC / flapping responses, fall back to solo
  per-PID polling instead of losing the whole batch.
- DTC multi-ECU merge window: merge DTC responses arriving from multiple
  ECUs within a collection window.
- 29-bit ISO 15765-4 end-to-end: scan probes 6 header/filter combos,
  extended-addressing flag persisted in NVS (obd_ext), RX filter set to
  ACCEPT_ALL when extended.
- Scan bus-alive gate; test_pid widened to uint16.
- Bump FIRMWARE_VERSION 1.1.25 -> 1.1.26.

Verified: native suite green (incl. new test_obd2_pid_table); bench-verified
on the S3 dash — clean boot, OTA reboot path clean (SW reset, no panic),
/api/selftest ok. OBD2 decode paths remain car-only unverified (no bench sim).

+++ main/net/web_server_signals.c
@@ -507,7 +507,7 @@ typedef struct {
     /* Test parameters (kept so _kick can call obd2_test_pid from the
      * LVGL task after lv_async_call hands off). */
     uint8_t           service;
-    uint8_t           pid;
+    uint16_t          pid;      /* 16-bit — Mode 22 PIDs exceed one byte */
     uint32_t          request_id;
     uint8_t           data_offset;
     uint8_t           data_bytes;
@@ -569,7 +569,7 @@ static esp_err_t _obd2_test_pid_handler(httpd_req_t *req)
     t = cJSON_GetObjectItemCaseSensitive(root, "service");
     ctx->service = cJSON_IsNumber(t) ? (uint8_t)t->valueint : 0x01;
     t = cJSON_GetObjectItemCaseSensitive(root, "pid");
-    ctx->pid = cJSON_IsNumber(t) ? (uint8_t)t->valueint : 0;
+    ctx->pid = cJSON_IsNumber(t) ? (uint16_t)t->valueint : 0;
     t = cJSON_GetObjectItemCaseSensitive(root, "request_id");
     ctx->request_id = cJSON_IsNumber(t) ? (uint32_t)t->valueint : 0;
     t = cJSON_GetObjectItemCaseSensitive(root, "data_offset");
